Produce a basic HTML version too.
This uses the HTML produced by default by the pandoc framework.
For now, there is very little tuning done on the output, just
a few CSS rules so that the table of contents is displayed on
the left and the main text next to that.
The CSS also asks for a sans serif font, which is just a personal
preference.

This also updates the name of the script to build the outputs, as
it no longer produces only PDF.

The github action workflow is adapted to also archive the HTML and
CSS files produced.