Automatically publish to github pages
This is done by running a github action step publishing the entire build
directory to the gh-pages branch. This only done for build actions that
are run on the main branch, and only when triggered by a push (and not
e.g. by a pull request).

One other minor addition is to make the book.html also available under
index.html, so that that's the page that gets picked up automatically on
the github pages webpage.