Automatically compute a version number.
This automatically computes a version number from the book, using the "git
describe" command.
"git describe --match '0'" looks for the tag with name "0", and computes a
string following the following pattern:
"tag"-"nr of commits since that tag"-g"current commit hash".

By having a tag with name "0" pointing to the first commit of the project's
history, this produces a version string looking like
0-"nr of commits since initial commit"-g"current commit hash".
For example: 0-56-gcebad2f.

Using this as a version string has the following advantages:
- The version name starts with "0" suggesting the book isn't complete yet.
- Since we maintain a linear history in this project, the second number
  will be a nice linearly increasing number, allowing humans to quickly
  which of two versions is more recent than the other.
- The commit hash at the end unambiguously points to the source code
  revision the book was build from; which can be useful for project
  contributors.

By further more adding '--dirty="-with-local-changes"' to the git descirbe
command line, we make sure that if there are local changes in the git checkout
of the project at build time, this becomes visible by appending
"-with-local-changes" to the version number.