Side-channel and covert channel chapter introduction (#75)
Side channels and covert channels are the basis for the attacks that make use
of them, which are described in later chapters.  Let's see if it makes sense to
separate out the description of covert channels from the actual attacks making
use of them. (I currently think it does make sense, but we probably will only
know when the chapters are far more complete).

If it turns out later that it's best to structure this content differently, we
can of course restructure it. For now, I think the restructuring in this commit
helps to unblock progress on writing content that should be in the book.

This commit only contains the introduction and a few needed section titles of
the covert channel and side-channel chapter. In the spirit of incremental
development, I think this is a useful increment to commit.