Make small/mobile screen view more readable
This commit adds some basic 'responsive design'
CSS to the default.css file such that the page
layout is minimally adjusted when the viewport
is less than 720px in width.

The main changes made for small screens are
moving the contents above the paragraph
elements and justifying paragraph text.

The advantage of the 'responsive design' approach
is that changes made for small screens have
no effect on the large screen styling.