Use bootstrap for HTML output
This is a first step towards moving to a much better HTML version of the
book, see discussion in issue #32.

In this commit, the bare minimum has been done to recreate the existing
layout with bootstrap. I've tried to add sufficient comments in the
pandoc_template.html file so that it should be clear what each line
does. I hope that will help with keeping the template file maintainable.

I see this is a first step on a longer road towards making a much
nicer-looking HTML output, aiming for something similar to what Rust's
mdbooks or bookdown.org books look like.

Note that by moving to bootstrap, the font seems to change a bit, and
there is more spacing between lines. Not sure if we should consider
trying to move back to the denser text from what we had before, or if we
should stick with bootstrap's default.

Also, I made the "header" screen-wide. I.e., the left-hand side TOC does
not appear at the top of the page, but under the header, i.e.  title,
author, copyright info. We can change that back later if wanted.

Otherwise, I believe that the look-and-feel and functionality of the
HTML output is the same as before.