[HTML] make sidebar toggle-able...
by clicking on the hamburger icon.
This is done using jquery javascript as there isn't a way (that
I found at least) doing this purely with bootstrap CSS.

This also removes the TOC modal as there is no need for it.

For smaller screen sizes, when a user clicks a link in the TOC,
it hides the TOC again as you can really only either look at the
TOC or at the content you want to jump to, not both at the same
time.