[HTML] main content: make scrollbar appear fully at right hand side.
Before this, the vertical scroll bar for the main content appeared right next
to the main content, and not further right, at the edge of the browser window.
Having the scroll bar at the edge of the browser window feels more natural. It
seems most, if not all, websites that show text with a maximum width do it this
way.