[HTML] Add sidenotes, converted from footnotes
This change adds a sidenotes margin on the right hand side.
It puts the footnotes in the right hand margin rather than at the bottom
of the document.

To make space for the sidenote marging, the TOC on the left hand side
only gets shown by default on xl screens.

To implement this required doing more things outside of the bootstrap
framework. As a result, we don't use the bootstrap framework for sizing
"columns" anymore. We still do use other features of the bootstrap
framework.

On smaller screens, the "sidenotes" get displayed inline, i.e. there is
a responsive design for this.