[HTML] Show \todo and \missingcontent as sidenote.
This shows todo's and missing contents as side notes in the HTML output.

For small screens, where sidenotes are shown inline, this commit does
not show the todos or missingcontents at all. It could be chosen later
to have an indication (small button?) that the reader could click to
make todos/missingcontent info visible.

We could also implement a button in the top row to toggle visibility of
todos/missingcontent.