[HTML] Make tooltips appear on the bottom of buttons
... rather than to the left, so that a tooltip is less likely to hide
other buttons.

However, for this first button ("toggle TOC"), this does not seem to be
working: the tooltip continues to be shown on the right hand side, i.e.
over the other buttons.

I could not find a way to change the behaviour.
Overall, this still is an improvement for the other buttons.