[HTML] Do not show tooltips on touch screen devices.
On touch screen devices, the tooltips are not really useful.
They only pop up after the user already pushes the button, so
they are then already "activating" the button action.

Therefore, disable them on touch devices.
The most consistent advice when looking for how to do that seems to
be to use "ontouchstart" in document.DocumentElement.
At least in Firefox this seems to work as expect across desktop and
mobile.