Improve mechanism to mark todo's and missingcontent
Before, we used \todo{} and \missingcontent{} to indicate todos in the
source, and also show that in the produced pdf and html.

The downside of using \todo{} and \missingcontent{} is that pandoc does
not interpret the text in the todos. As a result, it is not possible to
add e.g. links or any other formatting in the todo text.

This commit changes the mechanics of adding todos to using the idiomatic
pandoc Markdown way to do it: using pandoc 'fenced div's and 'bracketed
span's (see https://pandoc.org/MANUAL.html#divs-and-spans). This allows
pandoc to parse all text in todos just like any other text.

To still get the same output in the PDF version, this requires making
sure that the lua filter for handling todos also triggers when targeting
latex output.

This was implemented by refactoring the structure of the lua filters.
The pre-existing file "convert_to_sidenote.lua" now only handles
converting footnotes to sidenotes in html.

The new filter "markup_todo.lua" handles converting todo span and divs
into the appropriate representation for both the html and the latex
output.

This commit no longer makes a distinction between "\missingcontent" and
"\todo", they are both represented as todos.

An "inline" todo is written in the markdown source code using the
bracketed span syntax: [todo text]{.todo}. A "block" todo is written in
the markdown source code using the fenced div syntax:

::: TODO
todo text
:::