Convert .issue spands into links to github issues
This enables easily adding links to github issues in todos.
By having the links to github issues in todos, the hope is that readers
who would like to contribute their thoughts on how to implement a todo
will more likely do so.

As part of implementing this, I also discovered a bug in how the lua
script to markup todos was detecting which Spans to process. This is
also fixed in this commit.