Copy-edit: pedantry about NULL versus NUL.
'NULL' is the idiom in C for describing a null _pointer_: something
that is the size of a machine word and used to indicate the absence of
an address of memory elsewhere.

The term for the zero _character_ code, used to terminate strings in
C, is 'NUL' (as defined by ANSI X3.4).