Introduce transient execution attacks
This patch only introduces a high-level categorization of transient
execution attacks. The detailed description of the various Spectre-,
Meltdown-, LVI-style attacks will be for later patches.

Note that the table uses the pandoc grid table syntax, see
https://pandoc.org/chunkedhtml-demo/8.9-tables.html, section 8.9.4.
The version of pandoc we use in our docker image does not support that
yet. But it's the only variant of table that supports multi-row and
multi-column cells which I'd like to use for this table.
I'll need to figure out how to update the version of pandoc in the
docker container we use. That in itself shouldn't be too hard. The
annoying thing is that the pandoc filter pandoc-fignos that we use is no
longer maintained and does not work anymore with recent versions of
pandoc....
I'll look into update the version of pandoc we use in a later commit.