Update pandoc to recent version (3.0.1) so that we can use grid tables. (#225)
* Update pandoc to recent version (3.0.1) so that we can use grid tables.

pandoc-fignos is no longer supported on recent pandoc versions, so
instead, this commit adds the simple lua script theme/fignos.lua that
implements just the functionality that we are using. This results in the
figure numbering in the PDF to no use chapter-based numbering (e.g.
figure 2.3 for the 3rd figure in chapter 2), but rather uses linear
numbering (figures are numbered from 1 to N throughout the whole
document). This makes it consistent with how they are numbered in the
HTML output, so could be argued to actually be an improvement.

Diffing the tex output before and after this patch shows that the
only other difference is a minute difference in the links used for
citations, which is presumably caused by using the pandoc built-in
'citeproc' rather than the deprecated pandoc-citeproc filter.

The HTML output is harder to diff. The older version of pandoc creates
very long HTML lines, whereas the newer version of pandoc seemingly
tries hard to make the HTML output fit within an 80-column width.

As far as I can see, there are no regressions.
The only differences I've spotted are the ones described above.