Add an edit button after every header. (#227)
The intent is that this makes it easier for a reader to make a suggestion
for improvement when they read the book: they simple click one of the
"pencil" icons next to each header. The will bring them to the interactive
github interface to edit the source of the book near that header.