Improve and add detail to Control Flow Integrity section
I'm planning to add a better description of arm64e/pauthabi CFI, now that
support for it in upstream clang/llvm is getting completed.  I'm hoping to
explain aspects such as address diversity, signing and authentication oracles,
protecting vtable pointers etc.

When trying to write that, I found that some of the CFI concepts that apply to
multiple CFI schemes weren't covered well enough already in earlier sections,
to be able to explain these arm64e/pauthabi concepts easily.

Therefore, this patch is a major adaptation of the CFI section, expanding and
restructuring it quite a bit, in a way that hopefully:
1. will make it easier to understand for more readers, and
2. makes it easier to write an easy-to-understand expansion of the
   arm64e/pauthabi section in a follow-on patch.