Copy-edit: avoid using NULL to mean NUL. (#270)
'NULL', capitalized, is the idiom in C for describing a null
_pointer_: something that is the size of a machine word and used to
indicate the absence of an address of memory elsewhere. It shouldn't
bse used to refer to the zero _character_ code used to terminate
strings in C.

A very similar term which _does_ refer to the zero character code is
'NUL' (as defined in the ASCII specification, ANSI X3.4, along with
the rest of the C0 control codes like CR, LF, ESC etc). But that's now
archaic and obscure. Replaced with 'zero byte'.

Also added a link on "null-terminated string" for more information.