Clarify coarse- vs fine-grained CFI. (#269)
The introduction text explained a coarse-grained CFI scheme as having
"relatively coarse granularity", which is a non-explanation – if
someone is confused about what the "grain" metaphor means, it doesn't
help them to just repeat the same word again. Expanded it to spell out
that the "grains" in question are the equivalence classes of functions
that the same branch is legally allowed to target.